==> Experiment-1_Analysis/02_vis-and-posthoc.R <==
#################################
#        plots & post-hoc       #
#################################
# This script does:
# 1) post-hoc tests
# 2) accuracy and rt plots
# 3) model adjusted accuracy and rt plots
#
# saves plots to plots/ and respective experiment
# folders
# color = Experiment 1a
# grey = Experiment 1b
#
# Author: Aylin Kallmayer
# Year: 2022

if (!require("pacman")) install.packages("pacman")
p_load("dplyr","ggplot2","emmeans","remef","Rmisc","ggeffects","data.table")

# functions
source("src/01_mixed-models_functions.R")

#----Preprocess----
# run to preprocess new data
# source("00_preprocessing.R")

# update ggplot axis title theme
global_theme <- theme_classic()+
  theme(axis.title=element_text(size=12),
                      axis.text=element_text(size=10),
                      legend.position = "None")
#---------------------------------------------------------------------------------------------------#
#----EXPERIMENT 1A: color----
#---------------------------------------------------------------------------------------------------#

#----ACCURACY----
# with out badsrem data
data <- prepare_for_model(experiment="color", data_type="badsrem_outrem")
acc.model <- readRDS("models/color/acc-model_processed_badsrem_outrem.rds")
summary(acc.model)
mean(data$correct)
sd(data$correct)

#----Plotting non-canonical distributions----
agg_objs <- aggregate(data=data, correct~imagename+rotation+match, mean) %>%
  filter(match=="match")
(a <- ggplot(agg_objs, aes(x=as.factor(rotation), y=correct, color=match))+
    scale_color_manual(values = c("#999999", "#E69F00"))+
    geom_path(alpha=.3, aes(group=imagename))+
    geom_path(data=filter(agg_objs, imagename=="028_Wasserkocher_retroWasserkocher"),aes(group=1),size=1,color="orange",alpha=1)+
    geom_path(data=filter(agg_objs, imagename=="037_Mixer_Standmixer"),aes(group=1),size=1,color="blue",alpha=1)+
    geom_path(data=filter(agg_objs, imagename=="079_Tisch_Schreibtisch"),aes(group=1),size=1,color="purple",alpha=1)+
    xlab("Viewpoint")+
    ylab("Accuracy")+
    ylim(c(0,1))+
    global_theme)
# plot
png("plots/color/acc_obj_distr.png", units="cm", width=10, height=8, res=300)
a# insert ggplot code
dev.off()


#----Post-hoc tests----
# poly2 : match interaction
emtrends(acc.model, pairwise ~ match, var = "angle", max.degree = 2, adjust = "tukey")
# -> sig difference between match mismatch for quadratic degree
# -> quadratic trend only in match condition

#----Plot marginal effects with ggeffects----
(a <- ggpredict(acc.model, terms=c("angle [all]","match")) %>% ggplot(aes(x, predicted, color=group))+
  geom_point()+
  geom_path()+
  geom_ribbon(aes(ymin=conf.low, ymax=conf.high, fill=group), alpha=.3)+
  scale_color_manual(values = c("#999999", "#E69F00"))+
  scale_fill_manual(values = c("#999999", "#E69F00"))+
  scale_x_continuous(labels=c("1" = "0", "2" = "60", "3" = "120", "4" = "180", "5" = "240", "6" = "300"),
                     breaks = 1:6)+
  xlab("Viewpoint")+
  ylab("Adjusted Accuracy Predictions")+
  ggtitle("")+
  global_theme)
tiff("plots/color/acc_adjusted-badsrem_outrem.png", units="cm", width=10, height=8, res=300)
a# insert ggplot code
dev.off()

#----RT----
# with bads removed
data <- prepare_for_model(experiment="color",data_type="badsrem", RT=TRUE)
rt.model <- readRDS("models/color/rt-model_processed_badsrem.rds")
summary(rt.model)
mean(data$response_time)
sd(data$response_time)


#----Plot marginal effects with ggeffects----
(b <- ggpredict(rt.model, terms=c("angle [all]")) %>% ggplot(aes(x, predicted)) +
  geom_point()+
  geom_path()+
  geom_ribbon(aes(ymin=conf.low, ymax=conf.high), fill="grey", alpha=.3)+
  scale_x_continuous(labels=c("1" = "0", "2" = "60", "3" = "120", "4" = "180", "5" = "240", "6" = "300"),
                     breaks = 1:6)+
  xlab("Viewpoint")+
  ylab("Adjusted log RT Predictions")+
  ggtitle("")+
  global_theme)
tiff("plots/color/rt_adjusted-badsrem_outrem.png", units="cm", width=10, height=8, res=300)
b# insert ggplot code
dev.off()

#---------------------------------------------------------------------------------------------------#
#----EXPERIMENT 1B: GREY----
#---------------------------------------------------------------------------------------------------#  

#----ACCURACY----
# with out badsrem data
data <- prepare_for_model(experiment="grey",data_type="badsrem_outrem")
acc.model <- readRDS("models/grey/acc-model_processed_badsrem_outrem.rds")
summary(acc.model)
mean(data$correct)
sd(data$correct)

#----Plotting non-canonical distributions----
agg_objs <- aggregate(data=data, correct~imagename+rotation+match, mean) %>%
  filter(match=="match")

(a <- ggplot(agg_objs, aes(x=as.factor(rotation), y=correct, color=match))+
    scale_color_manual(values = c("#999999", "#E69F00"))+
    geom_path(alpha=.3, aes(group=imagename))+
    geom_path(data=filter(agg_objs, imagename=="028_Wasserkocher_retroWasserkocher"),aes(group=1),size=1,color="orange",alpha=1)+
    geom_path(data=filter(agg_objs, imagename=="037_Mixer_Standmixer"),aes(group=1),size=1,color="blue",alpha=1)+
    geom_path(data=filter(agg_objs, imagename=="079_Tisch_Schreibtisch"),aes(group=1),size=1,color="purple",alpha=1)+
    xlab("Viewpoint")+
    ylab("Accuracy")+
    ylim(c(0,1))+
    global_theme)
# plot
png("plots/grey/acc_obj_distr.png", units="cm", width=10, height=8, res=300)
a# insert ggplot code
dev.off()

# poly2 : match interaction
emtrends(acc.model, pairwise ~ match, var = "angle", max.degree = 2)
# -> sig difference between match mismatch for quadratic degree
# -> quadratic trend only in match condition

#----Plot marginal effects with ggeffects----
(c <- ggpredict(acc.model, terms=c("angle [all]","match")) %>% ggplot(aes(x, predicted, color=group))+
   geom_point()+
   geom_path()+
   geom_ribbon(aes(ymin=conf.low, ymax=conf.high, fill=group), alpha=.3)+
   scale_color_manual(values = c("#999999", "#E69F00"))+
   scale_fill_manual(values = c("#999999", "#E69F00"))+
   scale_x_continuous(labels=c("1" = "0", "2" = "60", "3" = "120", "4" = "180", "5" = "240", "6" = "300"),
                      breaks = 1:6)+
   xlab("Viewpoint")+
   ylab("Adjusted Accuracy Predictions")+
   ggtitle("")+
   global_theme)
tiff("plots/grey/acc_adjusted-badsrem_outrem.png", units="cm", width=10, height=8, res=300)
c# insert ggplot code
dev.off()

#----RT (log)----
# with bads removed
data <- prepare_for_model(experiment="grey",data_type="badsrem", RT=TRUE)
rt.model <- readRDS("models/grey/rt-model_processed_badsrem_log.rds")
summary(rt.model)
mean(data$response_time)
sd(data$response_time)

#----Plot marginal effects with ggeffects----
(d <- ggpredict(rt.model, terms=c("angle [all]")) %>% ggplot(aes(x, predicted)) +
   geom_point()+
   geom_path()+
   geom_ribbon(aes(ymin=conf.low, ymax=conf.high), fill="grey", alpha=.3)+
   scale_x_continuous(labels=c("1" = "0", "2" = "60", "3" = "120", "4" = "180", "5" = "240", "6" = "300"),
                      breaks = 1:6)+
   xlab("Viewpoint")+
   ylab("Adjusted log RT Predictions")+
   ggtitle("")+
   global_theme)
tiff("plots/grey/rt_adjusted-badsrem_outrem.png", units="cm", width=10, height=8, res=300)
d# insert ggplot code
dev.off()
